requirements.txt
asgiref==3.4.1
dj-places==5.1.0
Django==3.2
django-filter==21.1
django-location-field==2.1.0
djangorestframework==3.13.1
Markdown==3.3.6
Pillow==8.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
PyMySQL==1.0.2
